Refactor - moved VampireKnightToken to shared area
